refactor: hoist the shared bank transaction leaves to the domain root
This is the commit that decouples the three view variants. Every back-edge
between BankTransactionsTable, BankTransactionsList, and BankTransactionsMobileList
pointed at a shared leaf that happened to live inside one of them -- never at
another view. With the leaves hoisted, each of the three now has exactly one
consumer (BankTransactions.tsx), so they can move independently and this
migration needs no re-export shims.

Hoisted to features/bankTransactions/: BankTransactionsPaginatedList,
BankTransactionsSubmitButton, BankTransactionsAmountDate,
BankTransactionsListItemCategory, BankTransactionsSelectedValue.

BankTransactionsProcessingInfo came out of BankTransactionsList/ -- the table's
category cell and the mobile footer both render it, so living there is what made
that directory look circular.

Two renames:
- BankTransactionsTableEmptyState -> BankTransactionsDataStates. It exports
  BankTransactionsEmptyState and BankTransactionsErrorState, and the table and
  both lists use it, so the old name was wrong twice.
- BankTransactionRow/MatchBadge -> BankTransactionMatchBadge, deleting
  BankTransactionRow/ -- a directory named for a component it did not contain.
- BankTransactionsBaseSelectedValue -> BankTransactionsSelectedValue, so the
  directory namesake stops carrying a Base infix nothing mirrors.

CSS classes unchanged, including Layer__MatchTooltip in the renamed stylesheet.
