+++ src/components/BankTransactionsMobileList/BankTransactionsMobileList.tsx
@@ -10,12 +10,12 @@ import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFil
 import { VStack } from '@ui/Stack/Stack'
 import { MobileList } from '@blocks/MobileList/MobileList'
 import { useMobileListExpansion } from '@blocks/MobileList/useMobileListExpansion'
-import { BankTransactionsPaginatedList } from '@components/BankTransactions/BankTransactionsPaginatedList'
-import { BankTransactionsEmptyState, BankTransactionsErrorState } from '@components/BankTransactions/BankTransactionsTableEmptyState'
 import { BankTransactionsMobileBulkActionsHeader } from '@components/BankTransactionsMobileList/BankTransactionsMobileBulkActionsHeader'
 import { BankTransactionsMobileListItem } from '@components/BankTransactionsMobileList/BankTransactionsMobileListItem'
 import { BankTransactionsMobileListItemExpandedRow } from '@components/BankTransactionsMobileList/BankTransactionsMobileListItemExpandedRow'
 import { BankTransactionsMobileListItemFooter } from '@components/BankTransactionsMobileList/BankTransactionsMobileListItemFooter'
+import { BankTransactionsEmptyState, BankTransactionsErrorState } from '@features/bankTransactions/BankTransactionsDataStates/BankTransactionsDataStates'
+import { BankTransactionsPaginatedList } from '@features/bankTransactions/BankTransactionsPaginatedList/BankTransactionsPaginatedList'
 
 type BankTransactionsMobileListContentProps = {
   bankTransactions?: BankTransaction[]
