+++ src/components/MatchForm/MatchFormMobileItem.tsx
@@ -5,7 +5,7 @@ import { isMoneyIn } from '@utils/bankTransactions/shared'
 import { Checkbox } from '@ui/Checkbox/Checkbox'
 import { HStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
-import { BankTransactionsAmountDate } from '@components/BankTransactions/BankTransactionsAmountDate'
+import { BankTransactionsAmountDate } from '@features/bankTransactions/BankTransactionsAmountDate/BankTransactionsAmountDate'
 
 import './matchFormMobileItem.scss'
