+++ src/components/features/bankTransactions/BankTransactionsListItemCategory/BankTransactionsListItemCategory.tsx
@@ -2,9 +2,9 @@ import { useTranslation } from 'react-i18next'
 
 import { type BankTransaction } from '@internal-types/bankTransactions'
 import { useGetBankTransactionMatchOrCategoryWithDefault } from '@hooks/features/bankTransactions/useGetBankTransactionCategorizationWithDefault'
-import { BankTransactionsBaseSelectedValue } from '@components/BankTransactionsSelectedValue/BankTransactionsBaseSelectedValue'
-import { BankTransactionsCategorizedSelectedValue } from '@components/BankTransactionsSelectedValue/BankTransactionsCategorizedSelectedValue'
-import { BankTransactionsUncategorizedSelectedValue } from '@components/BankTransactionsSelectedValue/BankTransactionsUncategorizedSelectedValue'
+import { BankTransactionsCategorizedSelectedValue } from '@features/bankTransactions/BankTransactionsSelectedValue/BankTransactionsCategorizedSelectedValue'
+import { BankTransactionsSelectedValue } from '@features/bankTransactions/BankTransactionsSelectedValue/BankTransactionsSelectedValue'
+import { BankTransactionsUncategorizedSelectedValue } from '@features/bankTransactions/BankTransactionsSelectedValue/BankTransactionsUncategorizedSelectedValue'
 
 import './bankTransactionsListItemCategory.scss'
 
@@ -46,7 +46,7 @@ export const BankTransactionsListItemCategory = ({
       />
     )
     : (
-      <BankTransactionsBaseSelectedValue
+      <BankTransactionsSelectedValue
         type='placeholder'
         label={t('bankTransactions:empty.no_category_selected', 'No category selected')}
         className={className}
