+++ src/components/features/bankTransactions/BankTransactionsProcessingInfo/bankTransactionsProcessingInfo.scss
